Create Two_sum.cpp
This was a fun problem. The trick to the O(n) solution was to realize that to find the number that would needed to be summed with a given element is just (targert - arr[i]). If that number is also in the list and not at the same element then the target can be found.